requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.2
bleach==3.2.1
boto3==1.15.6
botocore==1.18.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dbfread==2.0.4
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
editdistance==0.5.3
entrypoints==0.3
fasttext==0.9.1
filelock==3.0.12
future==0.18.2
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.0
google-auth==1.22.0
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
ijson==2.6.1
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lmdb==0.98
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mdbtools==0.3.14
meza==0.42.5
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
nltk==3.4.5
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
omegaconf==2.0.1rc4
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycparser==2.20
Pygments==2.7.1
pygogo==0.13.2
pymongo==3.11.0
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==1.2.6
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.9.27
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
smmap==3.0.4
soupsieve==2.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toolz==0.11.1
torchtext==0.5.0
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.4
transformers==3.2.0
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
yarl==1.6.0
zipp==3.2.0